feat: supply routes

+++ src/service/supply.service.ts
@@ -5,14 +5,12 @@ import { AppError } from "../errors/AppError";
 import { ISupply } from "../interfaces/supply";
 
 export class SupplyService {
-  static async createSupplyService({
-    buy_price,
-    name,
-    provider_id,
-    order_id,
-  }: ISupply) {
+  static async create({ buy_price, name, provider_id, order_id }: ISupply) {
     const supplyRepository = AppDataSource.getRepository(Supply);
     const providerRepository = AppDataSource.getRepository(Providers);
+    if (!buy_price || !name || !provider_id || !order_id) {
+      throw new AppError(400, "Error in your request.");
+    }
 
     const provider = await providerRepository.findOne({
       where: { id: provider_id },
@@ -39,18 +37,27 @@ export class SupplyService {
   static readOne = async (supply_id: string) => {
     const supplyRepository = AppDataSource.getRepository(Supply);
     const supplyList = await supplyRepository.find();
+    if (!supply_id) {
+      throw new AppError(400, "Error in your request.");
+    }
 
     const supply = supplyList.find((supply) => supply.id === supply_id);
 
     return supply;
   };
 
-  static update = async ({ buy_price, name, supply_id }: ISupply) => {
+  static update = async (supply_id: string, { buy_price, name }: ISupply) => {
     const supplyRepository = AppDataSource.getRepository(Supply);
     const supplyList = await supplyRepository.find();
 
-    const supply = supplyList.find((supply) => supply.id === supply_id);
+    if (!supply_id || !buy_price || !name) {
+      throw new AppError(400, "Error in your request.");
+    }
+    const supply =  supplyList.find((supply) => supply.id === supply_id);
 
+    if (!supply) {
+      throw new AppError(404, "Supply not found.");
+    }
     await supplyRepository.update(supply!.id, { buy_price, name });
 
     return true;
@@ -61,7 +68,12 @@ export class SupplyService {
     const supplyList = await userRepository.find();
 
     const supply = supplyList.find((supply) => supply.id === id);
-
+    if (!id) {
+      throw new AppError(400, "Error in your request.");
+    }
+    if (!supply) {
+      throw new AppError(404, "Supply not found.");
+    }
     await userRepository.delete(supply!.id);
 
     return true;
